% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{available_genesets_goatrepo}
\alias{available_genesets_goatrepo}
\title{Discover available geneset collections from the GOAT GitHub repository}
\usage{
available_genesets_goatrepo()
}
\description{
Yields a table that describes available gene sets. You can download the actual data for the described versions with \code{download_genesets_goatrepo()}.
}
