% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{build_treemap}
\alias{build_treemap}
\title{Compute treemap data structures}
\usage{
build_treemap(ods, simplify = "none", toplevel_max_ngenes = Inf)
}
\arguments{
\item{ods}{result from \code{ontology_data_structures()}}

\item{simplify}{strategy for reducing the genesets returned in the treemap. Options;
"leaf_only" (most stringent, returns only leafs in the tree structure)
"prune_singletons" (remove parent terms that have exactly 1 child)
"pvalue" (remove parent terms where the child term p-value is at least 4 times better)
"none" (default; return all significant genesets that are not a "grouping term" in the treemap)}

\item{toplevel_max_ngenes}{groups in the treemap should not have more than this many genes ('ngenes' in geneset test results)}
}
\description{
\itemize{
\item shortlist = genesets/DAG-nodes to plot
\item update nested DAG data structure with recursive counts of shortlist elements
\item find level-1 elements to start with (possibly further down than direct children of root)
\item collapse children of level-1 elements (either from tree structure or from greedy aggregation starting at largest element)
\item restructure treemap data into format suitable for plotting
}
}
