% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dependency}
\alias{check_dependency}
\title{throw error if R package is unavailable}
\usage{
check_dependency(pkg, msg)
}
\arguments{
\item{pkg}{R package name}

\item{msg}{function name / reference for user}
}
\description{
throw error if R package is unavailable
}
