% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_similarity.R
\name{cluster_genesets}
\alias{cluster_genesets}
\title{cluster significant genesets from \code{test_genesets()} by geneset similarity (separately for each 'geneset source')}
\usage{
cluster_genesets(x, genelist, hclust_method = "ward.D2")
}
\arguments{
\item{x}{results from \code{test_genesets()}}

\item{genelist}{should be the same as provided to \code{test_genesets()}}

\item{hclust_method}{hierarchical clustering method, any of; 'ward.D', 'ward.D2' (default), 'single', 'complete', 'average'}
}
\value{
a list with elements genesets (param \code{x}), similarity, hc_row, hc_col
}
\description{
cluster significant genesets from \code{test_genesets()} by geneset similarity (separately for each 'geneset source')
}
