% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{darken_color}
\alias{darken_color}
\title{naively darken a color by mixing in black}
\usage{
darken_color(color, frac = 0.1)
}
\arguments{
\item{color}{input colors}

\item{frac}{fraction of black; >0 and <1}
}
\value{
adjusted value for input \code{color}
}
\description{
naively darken a color by mixing in black
}
