% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stats.R
\name{dsnorm}
\alias{dsnorm}
\title{Density function for skew normal distribution}
\usage{
dsnorm(x, mean = 0, sd = 1, xi = 1)
}
\arguments{
\item{x}{numeric vector of values}

\item{mean}{location parameter}

\item{sd}{scale parameter}

\item{xi}{skewness parameter}
}
\description{
Density function for skew normal distribution
}
