% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{edgelist_as_nested}
\alias{edgelist_as_nested}
\title{Convert edgelist table into a nested list structure}
\usage{
edgelist_as_nested(x, name, ngenes, edgelist)
}
\arguments{
\item{x}{element ID}

\item{name}{element name}

\item{ngenes}{element ngenes property}

\item{edgelist}{data.frame with properties parent, child, child_name, child_ngenes}
}
\description{
Convert edgelist table into a nested list structure
}
