% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{edgelist_find_children_recursive}
\alias{edgelist_find_children_recursive}
\title{depth-first recursion}
\usage{
edgelist_find_children_recursive(p, edgelist, result)
}
\arguments{
\item{p}{p}

\item{edgelist}{edgelist}

\item{result}{result}
}
\description{
depth-first recursion
}
