% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{edgelist_find_colored_parents_recursive}
\alias{edgelist_find_colored_parents_recursive}
\title{Within each recursive path; halt at first colored node}
\usage{
edgelist_find_colored_parents_recursive(x, geneset_table, query, result)
}
\arguments{
\item{x}{geneset ID}

\item{geneset_table}{data.frame with columns 'id' and 'parent_id'}

\item{query}{set of all colored nodes (vector)}

\item{result}{all colored parents, accumulated recursively (vector)}
}
\description{
Within each recursive path; halt at first colored node
}
