% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{edgelist_find_root}
\alias{edgelist_find_root}
\title{Find root ID and name in edgelist}
\usage{
edgelist_find_root(edgelist, genesets)
}
\arguments{
\item{edgelist}{result from e.g. \code{edgelist_from_ontology()}}

\item{genesets}{filtered genesets}
}
\description{
throw error if there is no single root
}
