% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stats.R
\name{fit_snorm_xi}
\alias{fit_snorm_xi}
\title{MLE estimate of parameter Xi in a skew-normal distribution}
\usage{
fit_snorm_xi(x, mu, sd, init_xi, lower = NA, upper = NA)
}
\arguments{
\item{x}{numeric vector}

\item{mu}{mean value for x}

\item{sd}{sd value for x}

\item{init_xi}{initial value for xi (typically 1~1.25)}

\item{lower}{lower bound for Xi}

\item{upper}{Upper bound for Xi}
}
\description{
MLE estimate of parameter Xi in a skew-normal distribution
}
