% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_similarity.R
\name{geneset_similarity_matrix}
\alias{geneset_similarity_matrix}
\title{geneset similarity matrix computation}
\usage{
geneset_similarity_matrix(
  mat,
  mat_row_geneid,
  mat_col_gsid,
  genelist_geneid,
  genelist_scores,
  weighted = TRUE
)
}
\arguments{
\item{mat}{numeric matrix}

\item{mat_row_geneid}{for mat rows; gene identifiers}

\item{mat_col_gsid}{for mat columns; geneset identifiers}

\item{genelist_geneid}{same identifiers as \code{mat_row_geneid}}

\item{genelist_scores}{a matrix where columns are various gene score types}

\item{weighted}{boolean value}
}
\description{
importantly, if multiple genelist_scores are provided, we use the gene score type for geneset j that yields the highest sum score
}
