% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_genesets.R
\name{genesets_as_matrix}
\alias{genesets_as_matrix}
\title{utility function that converts a genesets table into a sparse gene*geneset identity matrix}
\usage{
genesets_as_matrix(genesets)
}
\arguments{
\item{genesets}{genesets tibble. Must contain columns; "id", "genes", "ngenes"}
}
\description{
utility function that converts a genesets table into a sparse gene*geneset identity matrix
}
