% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets_go.R
\name{go_gene2go}
\alias{go_gene2go}
\title{parse gene2go file}
\usage{
go_gene2go(f, taxid_filter = 9606)
}
\arguments{
\item{f}{full path to gene2go file stored on the computer, e.g. previously downloaded from https://ftp.ncbi.nih.gov/gene/DATA/gene2go.gz}

\item{taxid_filter}{taxonomy id, integer}
}
\description{
note that it lacks parent/child relations, so from this file we only learn 'direct annotations'
}
