% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goat-package.R
\docType{package}
\name{goat}
\alias{goat-package}
\alias{goat}
\title{goat: Gene Set Analysis Using the Gene Set Ordinal Association Test}
\description{
Perform gene set enrichment analyses using the
Gene set Ordinal Association Test (GOAT) algorithm
and visualize your results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ftwkoopmans/goat/}
}

}
\author{
\strong{Maintainer}: Frank Koopmans \email{ftwkoopmans@gmail.com} (\href{https://orcid.org/0000-0002-4973-5732}{ORCID})

}
\keyword{internal}
