% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{goat_apply_precomputed_null}
\alias{goat_apply_precomputed_null}
\title{Compute the parameters of a skew-normal distribution, given a previously fitted polynomial function over geneset sizes}
\usage{
goat_apply_precomputed_null(usize, par)
}
\arguments{
\item{usize}{integer vector of geneset sizes}

\item{par}{result from \code{goat_nulldistribution_function}}
}
\value{
data.frame with columns; size, mu, sigma, xi
}
\description{
Compute the parameters of a skew-normal distribution, given a previously fitted polynomial function over geneset sizes
}
