% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goat_example_datasets}
\alias{goat_example_datasets}
\title{Previously published datasets that were used as examples in the GOAT manuscript}
\format{
\subsection{\code{goat_example_datasets}}{

a list of data.frames that represent differential expression analysis from high-throughput datasets
}
}
\source{
The name of each dataset in \code{goat_example_datasets} contains the respective PubMed identifier
}
\usage{
goat_example_datasets
}
\description{
Previously published datasets that were used as examples in the GOAT manuscript
}
\keyword{datasets}
