% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goat_nulldistributions.R
\name{goat_nulldistribution_function}
\alias{goat_nulldistribution_function}
\title{Estimate null distribution parameters for a given genelist length as a polynomial function of geneset sizes}
\usage{
goat_nulldistribution_function(
  genelist_N,
  niter = 5e+05,
  return_fit_objects = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{genelist_N}{genelist length. Integer value between 100 and 50000}

\item{niter}{number of bootstrap iterations that should be performed for generating empirical null distributions. Note that setting a large number will dramatically increase the RAM usage of this function! Integer value between 10000 and 5000000}

\item{return_fit_objects}{should the 'fit' object, which can be used as input for \code{predict}, be included in results? boolean value, TRUE or FALSE, default; FALSE}

\item{verbose}{should plots be created that describe the accuracy of fitting sd and xi? boolean value, TRUE or FALSE, default; FALSE}
}
\description{
\enumerate{
\item For a range of geneset sizes, from smallest (10) to largest (genelist length / 2), generate permutation-based null distributions and estimate their skew-normal parameters mean, sd, xi
\item Fit monotonic polynomial functions across geneset sizes that predict sd or xi
}
}
