% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goat_nulldistributions.R
\name{goat_nulldistribution_independent}
\alias{goat_nulldistribution_independent}
\title{Estimate null distribution parameters for a given set of geneset sizes and an array of gene scores (from genelist)}
\usage{
goat_nulldistribution_independent(geneset_sizes, gene_scores, niter, verbose)
}
\arguments{
\item{geneset_sizes}{vector of geneset sizes; integer values larger than 10 and smaller than half the genelist length (gene_scores length)}

\item{gene_scores}{gene score vector; these values are assumed to be the 'universe' of all possible gene-level scores. These are used to compute geneset score null distributions}

\item{niter}{number of bootstrap iterations that should be performed for generating empirical null distributions. Note that setting a large number will dramatically increase the RAM usage of this function! Integer value between 10000 and 5000000}

\item{verbose}{boolean, plot 10 null distribution histograms augmented with fitted skew-normal distributions}
}
\value{
data.frame with columns; genelist_N, size, mu, sigma, xi
}
\description{
Estimate null distribution parameters for a given set of geneset sizes and an array of gene scores (from genelist)
}
