% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goat_nulldistributions}
\alias{goat_nulldistributions}
\title{Precomputed parameters used by the GOAT algorithm}
\format{
\subsection{\code{goat_nulldistributions}}{

a data.frame with precomputed GOAT null distribution parameters
}
}
\usage{
goat_nulldistributions
}
\description{
there parameters are used by goat to efficiently perform geneset testing without bootstrapping
}
\keyword{datasets}
