% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{goat_testgene_score_matrix}
\alias{goat_testgene_score_matrix}
\title{From effectsize and pvalue in the genelist table to a matrix of 'rank transformed gene scores'}
\usage{
goat_testgene_score_matrix(genelist, score_type)
}
\arguments{
\item{genelist}{see \code{test_genesets_goat_precomputed}}

\item{score_type}{see \code{test_genesets_goat_precomputed}}
}
\value{
gene*score_type matrix where rows are aligned with the rows of the genelist table
}
\description{
Depending on the score_type, we sort the genelist by pvalue or effectsize (up/down/absolute) and
then compute gene scores. Note that it is common for genelists to have many genes with \code{pvalue=1},
which result in ties when sorting. We break ties for pvalue sorting by absolute effectsizes
(if available) and we break ties for (up/down/absolute) effectsizes by pvalues (if available).
Finally, we break any remaining ties by sorting the gene column to ensure stable sorting behaviour
under all circumstances.

Technical note; after upstream input validation, we're assured this will yield a non-empty matrix with at least 1 column
}
