% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{goat_version}
\alias{goat_version}
\title{Return goat package version as a string}
\usage{
goat_version()
}
\value{
package version as a string
}
\description{
simple wrapper around utils::packageVersion()
}
