% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{hgnc_idmap_table}
\alias{hgnc_idmap_table}
\title{Parse HGNC gene identifier lookup table that was downloaded from genenames.org into a table with HGNC ID, symbol, synonym (NA if unavailable), entrez ID}
\usage{
hgnc_idmap_table(filename)
}
\arguments{
\item{filename}{full path to the downloaded table (expected to be tsv format, typically has .txt or .tsv file extension)}
}
\value{
a long-format table with columns; hgnc_id, hgnc_symbol, type, value
}
\description{
download link: https://www.genenames.org/download/statistics-and-files/
table: "Complete dataset download links" -->> "Complete HGNC approved dataset" -->> download the "TXT" table
filename is typically something like hgnc_complete_set.txt
URL; https://ftp.ebi.ac.uk/pub/databases/genenames/hgnc/tsv/hgnc_complete_set.txt
}
\section{alternatively;}{
table: "Total Approved Symbols" -->> "TXT" / "text file in TSV format"
filename is typically something like non_alt_loci_set.txt
}

