% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lighten_color}
\alias{lighten_color}
\title{naively lighten a color by mixing in white}
\usage{
lighten_color(color, frac = 0.1)
}
\arguments{
\item{color}{input colors}

\item{frac}{fraction of white; >0 and <1}
}
\description{
naively lighten a color by mixing in white
}
