% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_genesets.R
\name{methods_text}
\alias{methods_text}
\title{create M&M text}
\usage{
methods_text(x, genelist, settings)
}
\arguments{
\item{x}{see \code{save_genesets()}}

\item{genelist}{see \code{save_genesets()}}

\item{settings}{settings attribute attached to the genesets result table}
}
\description{
create M&M text
}
