% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{nested_aggregate_child_stats}
\alias{nested_aggregate_child_stats}
\title{For each element in a DAG, count the number of recursive child terms and those that are on shortlist}
\usage{
nested_aggregate_child_stats(x, shortlist)
}
\arguments{
\item{x}{list representing a nested DAG}

\item{shortlist}{vector of element IDs}
}
\description{
For each element in a DAG, count the number of recursive child terms and those that are on shortlist
}
