% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{nested_find_equivalent_child_recursive}
\alias{nested_find_equivalent_child_recursive}
\title{Recursively replace a DAG element by its children while the child has the same gene count as the parent}
\usage{
nested_find_equivalent_child_recursive(x)
}
\arguments{
\item{x}{list representing a nested DAG}
}
\description{
Recursively replace a DAG element by its children while the child has the same gene count as the parent
}
