% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{ontology_data_structures}
\alias{ontology_data_structures}
\title{Compute tree and treemap data structures from geneset tables that include parent/child links}
\usage{
ontology_data_structures(geneset_ids, genesets, genesets_test_result)
}
\arguments{
\item{geneset_ids}{vector of geneset identifiers}

\item{genesets}{entire geneset table; typically the complete GO database}

\item{genesets_test_result}{geneset testing results; the output from \code{test_genesets()}}
}
\description{
Compute tree and treemap data structures from geneset tables that include parent/child links
}
