% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_genes.R
\name{partition_genes}
\alias{partition_genes}
\title{Classify genes into 2 groups, e.g. to define significant or topN genes, resulting in a 'signif' column with boolean values}
\usage{
partition_genes(
  genes,
  col,
  decreasing = FALSE,
  use_abs = FALSE,
  cutoff = NULL,
  fraction = NULL,
  topn = NULL
)
}
\arguments{
\item{genes}{gene tibble where each row is a unique gene, must contain column name \code{col}}

\item{col}{column name in \code{genes}}

\item{decreasing}{order \code{col} in descending (set TRUE) or ascending order (set FALSE, default) prior to partitioning?}

\item{use_abs}{use absolute values (default FALSE), e.g. when setting a threshold on effect-sizes}

\item{cutoff}{threshold for values in \code{col} to select  (must provide exactly 1 parameter for filtering, either \code{cutoff}, \code{fraction} or \code{topn})}

\item{fraction}{fraction of rows in \code{genes} tibble to select  (must provide exactly 1 parameter for filtering, either \code{cutoff}, \code{fraction} or \code{topn})}

\item{topn}{number of rows in \code{genes} tibble to select  (must provide exactly 1 parameter for filtering, either \code{cutoff}, \code{fraction} or \code{topn})}
}
\description{
This can be convenient to prepare the significant/test/foreground set for classical ORA,
e.g. \code{test_genesets()} with parameter \code{method = "fisherexact"}.
}
\examples{
\dontrun{
  # significant hits
  genelist = partition_genes(genelist, col="pvalue_adjust", decreasing=F, cutoff=0.01)
  # abs(effectsize) >= 4
  genelist = partition_genes(genelist, col="effectsize", decreasing=T, use_abs=T, cutoff=4)
  # top 10\% 'best' p-values
  genelist = partition_genes(genelist, col="pvalue", decreasing=F, fraction = 0.1)
}
}
