% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_similarity.R
\name{plot_network}
\alias{plot_network}
\title{plot geneset distance matrix as a network}
\usage{
plot_network(
  clusters,
  src,
  show_clusters = TRUE,
  show_text = FALSE,
  topn_edges = 5,
  clr_default = "#29b6f6",
  node_color_palette = goat::gg_color_hue
)
}
\arguments{
\item{clusters}{result from \code{cluster_genesets}}

\item{src}{source property (e.g. "GO_CC")}

\item{show_clusters}{boolean value}

\item{show_text}{boolean value}

\item{topn_edges}{topN edges to retain per geneset (typically 5~8)}

\item{clr_default}{default color for the network, used only when \code{show_clusters} is set to \code{FALSE}}

\item{node_color_palette}{function with 1 parameter, N, that returns N colors (default is goat function \code{gg_color_hue})}
}
\value{
a ggplot2 object
}
\description{
plot geneset distance matrix as a network
}
