% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stats.R
\name{psnorm_upper_tail}
\alias{psnorm_upper_tail}
\title{Probability function for skew normal distribution}
\usage{
psnorm_upper_tail(q, mean = 0, sd = 1, xi = 1)
}
\arguments{
\item{q}{numeric vector of quantiles}

\item{mean}{location parameter}

\item{sd}{scale parameter}

\item{xi}{skewness parameter}
}
\description{
Probability function for skew normal distribution
}
