% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stats.R
\name{rankscore_fixed_order}
\alias{rankscore_fixed_order}
\title{mock data on same scale as \code{rankscore}}
\usage{
rankscore_fixed_order(n)
}
\arguments{
\item{n}{genelist length}
}
\description{
mock data on same scale as \code{rankscore}
}
