% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_trunc_right}
\alias{string_trunc_right}
\title{simple string truncation}
\usage{
string_trunc_right(string, width, trim_left = FALSE)
}
\arguments{
\item{string}{string that should be truncated}

\item{width}{desired max length}

\item{trim_left}{instead of right trunc (default), do left instead}
}
\value{
truncated variant of input \code{string}
}
\description{
replacement for stringr::trunc() so we don't need a package dependency for just 1 function (our code was adapter therefrom)
}
