% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{symbol_to_entrez}
\alias{symbol_to_entrez}
\title{Map the the symbol column in a table to HGNC human gene IDs by matching official gene symbols and synonyms}
\usage{
symbol_to_entrez(x, hgnc)
}
\arguments{
\item{x}{a data.table with a column symbol}

\item{hgnc}{HGNC lookup table from \code{hgnc_lookuptable()}}
}
\description{
Map the the symbol column in a table to HGNC human gene IDs by matching official gene symbols and synonyms
}
\examples{
\dontrun{
  # TODO: update the filename to your downloaded file
  # download instructions @ `hgnc_lookuptable`
  f = "C:/data/hgnc_complete_set.txt"
  df = data.frame(symbol = c("vamp2", "STXBP1", "UNC18", NA, "PSD95", "NOT-A-GENE"))
  hgnc = hgnc_lookuptable(f)
  symbol_to_entrez(df, hgnc)
}
}
