% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{test_genesets_goat}
\alias{test_genesets_goat}
\title{Implementation of the GOAT algorithm}
\usage{
test_genesets_goat(genesets, genelist, score_type, nulldistribution_parameters)
}
\arguments{
\item{genesets}{see \code{test_genesets_goat_precomputed}}

\item{genelist}{see \code{test_genesets_goat_precomputed}}

\item{score_type}{see \code{test_genesets_goat_precomputed}}

\item{nulldistribution_parameters}{result from goat_nulldistribution_independent}
}
\description{
Implementation of the GOAT algorithm
}
