% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_genelist.R
\name{validate_genelist}
\alias{validate_genelist}
\title{validate a table with genes (that should be tested in overrepresentation-analysis) for compatibility with this R package}
\usage{
validate_genelist(genelist)
}
\arguments{
\item{genelist}{gene tibble to validate}
}
\description{
validate a table with genes (that should be tested in overrepresentation-analysis) for compatibility with this R package
}
