% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{validate_goat_genelist}
\alias{validate_goat_genelist}
\title{helper function to validate input parameters for GOAT computation functions}
\usage{
validate_goat_genelist(x, score_type)
}
\arguments{
\item{x}{object to validate}

\item{score_type}{score type}
}
\description{
helper function to validate input parameters for GOAT computation functions
}
