% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{validate_goat_genesets}
\alias{validate_goat_genesets}
\title{helper function to validate input parameters for GOAT computation functions}
\usage{
validate_goat_genesets(x, genelist_N)
}
\arguments{
\item{x}{object to validate}

\item{genelist_N}{genelist length}
}
\description{
helper function to validate input parameters for GOAT computation functions
}
