% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{validate_goat_niter}
\alias{validate_goat_niter}
\title{helper function to validate input parameters for GOAT computation functions}
\usage{
validate_goat_niter(x)
}
\arguments{
\item{x}{object to validate}
}
\description{
helper function to validate input parameters for GOAT computation functions
}
