% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goat_nulldistributions}
\alias{goat_nulldistributions}
\title{Precomputed parameters used by the GOAT algorithm}
\format{
\subsection{\code{goat_nulldistributions}}{

precomputed null distribution parameters, used by goat to efficiently perform geneset testing without bootstrapping
}
}
\usage{
goat_nulldistributions
}
\description{
Precomputed parameters used by the GOAT algorithm
}
\keyword{datasets}
