% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{goat_print_version}
\alias{goat_print_version}
\title{Print package version and logo to console}
\usage{
goat_print_version()
}
\description{
Print package version and logo to console
}
\examples{
goat_print_version()
}
