% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets_go.R
\name{load_genesets_go_fromfile}
\alias{load_genesets_go_fromfile}
\title{construct a geneset table from gene2go and OBO files}
\usage{
load_genesets_go_fromfile(
  file_gene2go,
  file_goobo,
  include_child_annotations = TRUE
)
}
\arguments{
\item{file_gene2go}{gene2go file from NCBI. Also works with the gzipped file gene2go.gz}

\item{file_goobo}{OBO file from geneontology.org}

\item{include_child_annotations}{boolean; include annotations against child terms? In most situations, TRUE (default) is the desired setting}
}
\value{
table with columns; source (character), source_version (character), id (character), name (character), genes (list), ngenes (int)
}
\description{
Download link for gene2go file; https://ftp.ncbi.nih.gov/gene/DATA/gene2go.gz
Download link for gene ontology OBO file; http://current.geneontology.org/ontology/go.obo
}
\examples{
\dontrun{
  genesets_asis = load_genesets_go_fromfile(
    file_gene2go = "C:/DATA/download_2023-01-01/gene2go",
    file_goobo = "C:/DATA/download_2023-01-01/go.obo"
  )
}
}
