% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_genesets.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot the geneset similarity matrix as a heatmap}
\usage{
plot_heatmap(
  x,
  output_dir = NA,
  heatmap_colors = grDevices::hcl.colors(100, "Viridis", rev = F),
  fontsize = 10
)
}
\arguments{
\item{x}{result from \code{cluster_genesets()}}

\item{output_dir}{full path to output directory. Set to NA to directly show the figures instead of writing them to file}

\item{heatmap_colors}{a vector of 100 colors to be used for the heatmap (101 breaks are computed between 0 and the max value in the distance matrix)}

\item{fontsize}{parameter sent to pheatmap::pheatmap(); control the size of labels in the plot, defaults to 10. Note that you can also change the plot device size, see examples}
}
\description{
plot the geneset similarity matrix as a heatmap
}
\examples{
\dontrun{
  # try various color palettes
  plot_heatmap(plot_heatmapclusters, heatmap_colors = hcl.colors(100, "Viridis", rev = F))
  plot_heatmap(plot_heatmapclusters, heatmap_colors = hcl.colors(100, "Inferno", rev = F))
  plot_heatmap(plot_heatmapclusters, heatmap_colors = hcl.colors(100, "Lajolla", rev = T))
  plot_heatmap(plot_heatmapclusters, heatmap_colors = hcl.colors(100, "Mako", rev = F))
  plot_heatmap(plot_heatmapclusters, heatmap_colors = hcl.colors(100, "Turku", rev = T))
  plot_heatmap(plot_heatmapclusters, heatmap_colors = hcl.colors(100, "Grays", rev = T))
}
}
