% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_genesets.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Plot a lollipop visualization of genesets}
\usage{
plot_lollipop(
  x,
  output_dir = NA,
  only_reduced = FALSE,
  show_pvalue = FALSE,
  score = "minlogp",
  score_colorscale = "source",
  max_ngenes = NA,
  topn = NA,
  padj_cutoff = NA
)
}
\arguments{
\item{x}{results from function \code{test_genesets}}

\item{output_dir}{full path to output directory. Set to NA to directly show the figures instead of writing them to file}

\item{only_reduced}{only show the reduced/summarized set of significant genesets. This requires that you first applied the \code{reduce_genesets} function}

\item{show_pvalue}{boolean parameter that indicates whether adjusted p-values should be shown next to each data point}

\item{score}{type of score to show on the x-axis. Options: "minlogp" to show -log10(adjusted pvalue), which is default. Set "oddsratio" to show the enrichment of significant genes.
For further details on this score and its computation, see the \code{score_geneset_oddsratio} function documentation.
Basically, the genesets in this plot are sorted by their proportion of foreground/significant genes
(and this ratio is standardized against the overall ratio of significant genes as to make this statistic comparable across analyses/datasets).}

\item{score_colorscale}{defines the limits for the color scales. options; \code{score_colorscale = "source"} use the range of values per 'source' to compute colors scales (default). Set to "overall" in order to have a unified color scale across 'source'  (e.g. same color bar across GO_CC/GO_MF/GO_BP). Alternatively, provide a numeric vector of 2 values to manually define lower/upper-limits.}

\item{max_ngenes}{only plot terms with less than N genes (quick way to get rid of large/unspecific terms)}

\item{topn}{topn terms to show after sorting genesets by p-value. For example, this makes it easy to plot the top10 GO terms. Set to NA to ignore this filter (default)}

\item{padj_cutoff}{adjusted pvalue cutoff for terms shown in the plot. If set to NA (default), all significant genesets are used (i.e. 'signif' column in the input geneset table)}
}
\description{
Plot a lollipop visualization of genesets
}
