% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genesets_goat.R
\name{test_genesets_goat_precomputed}
\alias{test_genesets_goat_precomputed}
\title{Test geneset enrichment with the Geneset Ordinal Association Test (GOAT) algorithm}
\usage{
test_genesets_goat_precomputed(genesets, genelist, score_type)
}
\arguments{
\item{genesets}{genesets data.frame, must contain columns; "source", "id", "genes", "ngenes"}

\item{genelist}{genelist data.frame, must contain columns "gene" and "pvalue"/"effectsize" (depending on parameter \code{score_type})}

\item{score_type}{how to compute gene scores?
Option "pvalue" uses values from the pvalue column in \code{genelist} in a one-way test for enrichment; lower p-value is better
Option "effectsize" uses values from the effectsize column in \code{genelist} in a two-way test for enrichment; is a geneset enriched in either down- or up-regulated genes?
Option "effectsize_abs" uses values from the effectsize column in \code{genelist} in a one-way test for enrichment; is a geneset enriched when testing absolute effectsizes?
Option "effectsize_up" uses values from the effectsize column in \code{genelist} in a one-way test for enrichment; is a geneset enriched in up-regulated genes?  (i.e. positive effectsize)
Option "effectsize_down" uses values from the effectsize column in \code{genelist} in a one-way test for enrichment; is a geneset enriched in down-regulated genes?  (i.e. negative effectsize)}
}
\description{
In most cases, it's more convenient to call the more generic \code{test_genesets} function which also applies multiple-testing correction (per geneset source) to the geneset p-values computed by this function.

This is the canonical geneset test function for GOAT that uses precomputed null distributions that are bundled with the GOAT package
}
\seealso{
\code{test_genesets}
}
