% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ontology.R
\name{treemap_plot}
\alias{treemap_plot}
\title{Plot a treemap}
\usage{
treemap_plot(x, label_group = FALSE)
}
\arguments{
\item{x}{\code{treemap_plotdata} data table that was computed by the \code{treemap_data()} function}

\item{label_group}{set TRUE to show only group-level labels}
}
\description{
simple wrapper around the treemap R package. To customize this plot, copy/paste its code and tweak parameters as desired
}
\examples{
\dontrun{
# use GOAT as per usual
genelist = goat_example_datasets[["Hondius 2021:mass-spec:PMID33492460"]]
genesets = load_genesets_go_bioconductor()
genesets_filtered = filter_genesets(genesets, genelist)
goat_result = test_genesets(
  genesets_filtered, genelist, method = "goat", score_type = "effectsize",
  padj_method = "bonferroni", padj_cutoff = 0.05
)
# subset GO CC results
x = goat_result |> filter(signif & source == "GO_CC")
tm = treemap_data(
  geneset_ids = x$id,
  genesets = genesets,
  genesets_test_result = x,
  simplify = "leaf_only" # options: none/leaf_only/prune_singletons/pvalue
)
treemap_plot(tm$treemap_plotdata)
}
}
